+++ src/app/compras/actions/orders.ts
@@ -14,52 +14,3 @@ export async function listOrdersAction(): Promise<PurchaseOrder[]> {
   return mockOrders;
 }
 
-// ========= CREAR =========
-export async function createOrderAction(payload: {
-  proveedorId: string;
-  deposito: string;
-  fechaEntrega: string; // yyyy-mm-dd
-  items: PurchaseOrderItem[];
-}): Promise<PurchaseOrder> {
-  // En real: POST a tu API, obtendrás el objeto creado
-  // Demo con mocks:
-  const newId = genOCId(mockOrders);
-  const now = new Date();
-  const creationDate = toDDMMYYYY(`${now.getFullYear()}-${pad(now.getMonth() + 1)}-${pad(now.getDate())}`);
-  const creationTime = `${pad(now.getHours())}:${pad(now.getMinutes())}`;
-
-  const supplier = mockSuppliers.find((s) => s.id === payload.proveedorId)!;
-  const totalCantidad = payload.items.reduce((s, i) => s + i.quantity, 0);
-  const totalMonto = payload.items.reduce((s, i) => s + i.totalPrice, 0);
-
-  const created: PurchaseOrder = {
-    id: newId,
-    creationDate,
-    creationTime,
-    supplier: { id: supplier.id, name: supplier.name, code: supplier.code },
-    warehouse: payload.deposito,
-    deliveryDate: toDDMMYYYY(payload.fechaEntrega),
-    status: "Incompleta",
-    total: totalMonto,
-    totalQuantity: totalCantidad,
-    items: payload.items,
-  };
-
-  // En mocks no persistimos; con API real persistís y revalidás:
-  revalidatePath("/compras");
-  return created;
-}
-
-
-
-// Helpers locales
-function pad(n: number) { return n.toString().padStart(2, "0"); }
-function toDDMMYYYY(yyyy_mm_dd: string) {
-  const [y, m, d] = yyyy_mm_dd.split("-").map(Number);
-  return `${pad(d)}/${pad(m)}/${y}`;
-}
-function genOCId(all: PurchaseOrder[]) {
-  const nums = all.map((o) => Number(o.id.split("-")[1] || "0")).filter((n) => !isNaN(n));
-  const next = (nums.length ? Math.max(...nums) : 0) + 1;
-  return `OC-${String(next).padStart(3, "0")}`;
-}
